Making suggested changes
