--- browser/frontend/lambdas/requirements-dev.txt
@@ -1,1 +1,0 @@
--r ../../../requirements-dev.txt
