+++ browser/lambdas/chalice/cloudfront-invalidator/.chalice/config.in.json
