+++ browser/lambdas/iam/policy-templates/cloudfront-invalidator-lambda.json
